hyb basic samples : change int to double in hello world sample

+++ HybridizerBasicSamples/1.Simple/HelloWorld/HelloWorld/Program.cs
@@ -7,15 +7,15 @@ namespace HelloWorld
     class Program
     {
         [EntryPoint("run")]
-        public static void Run(int N, int[] a, int[] b)
+        public static void Run(int N, double[] a, double[] b)
         {
             Parallel.For(0, N, i => { a[i] += b[i]; });
         }
 
         static void Main(string[] args)
         {
-            int[] a = { 1, 2, 3, 4, 5 };
-            int[] b = { 10, 20, 30, 40, 50 };
+            double[] a = { 1, 2, 3, 4, 5 };
+            double[] b = { 10, 20, 30, 40, 50 };
 
             // create an instance of HybRunner object to wrap calls on GPU
             HybRunner runner = HybRunner.Cuda("HelloWorld_CUDA.dll");
